Moved Logger into it's own file
